+++ src/index.html
@@ -15,7 +15,9 @@
             <div class="title">
                 <h1>TV Hub</h1>
             </div>
-
+            <div class="pages"><a href="">One</a>
+                <a href="">Two</a>
+            </div>
 
         </nav>
     </header>
